requirements.txt
fitz
tqdm
python-docx
Pillow
openai
pypdf
pymupdf
PyPDF2  
